+++ packages/server/controllers/apps.test.js
@@ -71,6 +71,8 @@ describe('app controller', () => {
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -79,6 +81,8 @@ describe('app controller', () => {
       {
         path: 'another-app',
         definition: { name: 'Another App', defaultPage: 'Another Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -122,6 +126,8 @@ defaultPage: Another Page
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -131,6 +137,8 @@ defaultPage: Another Page
         path: 'another-app',
         private: true,
         definition: { name: 'Another App', defaultPage: 'Another Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -166,6 +174,8 @@ defaultPage: Test Page
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -193,6 +203,8 @@ defaultPage: Test Page
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -203,6 +215,8 @@ defaultPage: Test Page
       {
         path: 'test-app-b',
         definition: { name: 'Test App B', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationB.id,
       },
       { raw: true },
@@ -567,6 +581,8 @@ pages:
           {
             path: index + 1 === 1 ? 'test-app' : `test-app-${index + 1}`,
             definition: { name: 'Test App', defaultPage: 'Test Page' },
+            vapidPublicKey: `a${index}`,
+            vapidPrivateKey: `b${index}`,
             OrganizationId: organizationId,
           },
           { raw: true },
@@ -760,6 +776,8 @@ pages:
       {
         definition: { name: 'Test App', defaultPage: 'Test Page' },
         path: 'test-app',
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -818,6 +836,8 @@ pages:
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -863,6 +883,8 @@ pages:
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -908,6 +930,8 @@ pages:
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -949,6 +973,8 @@ pages:
       {
         path: 'foo',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -968,6 +994,8 @@ pages:
       {
         path: 'foo',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -987,6 +1015,8 @@ pages:
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -1031,6 +1061,8 @@ pages:
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: newOrganization.id,
       },
       { raw: true },
@@ -1074,6 +1106,8 @@ pages:
       {
         path: 'foo',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -1081,7 +1115,7 @@ pages:
     const response = await request(server)
       .patch(`/api/apps/${appA.id}`)
       .set('Authorization', token)
-      .field('definition', JSON.stringify({ name: 'Foobar' }));
+      .field('app', JSON.stringify({ name: 'Foobar' }));
 
     expect(response.status).toBe(400);
   });
@@ -1133,6 +1167,8 @@ pages:
       {
         path: 'test-app',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organization.id,
       },
       { raw: true },
@@ -1150,6 +1186,8 @@ pages:
       {
         path: 'bar',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -1195,6 +1233,8 @@ pages:
       {
         path: 'bar',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -1253,6 +1293,8 @@ pages:
       {
         path: 'bar',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -1316,6 +1358,8 @@ pages:
       {
         path: 'bar',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -1356,6 +1400,8 @@ pages:
       path: 'b',
       private: false,
       definition: { name: 'Test App', defaultPage: 'Test Page' },
+      vapidPublicKey: 'a',
+      vapidPrivateKey: 'b',
       OrganizationId: organizationId,
     });
 
@@ -1397,6 +1443,8 @@ pages:
       {
         path: 'bar',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },
@@ -1422,6 +1470,8 @@ pages:
       {
         path: 'bar',
         definition: { name: 'Test App', defaultPage: 'Test Page' },
+        vapidPublicKey: 'a',
+        vapidPrivateKey: 'b',
         OrganizationId: organizationId,
       },
       { raw: true },