+++ packages/server/controllers/templates.js
@@ -2,6 +2,7 @@ import { normalize } from '@appsemble/utils';
 import Boom from '@hapi/boom';
 import crypto from 'crypto';
 import { UniqueConstraintError } from 'sequelize';
+import { generateVAPIDKeys } from 'web-push';
 
 import templates from '../templates/apps';
 import getAppFromRecord from '../utils/getAppFromRecord';
@@ -38,6 +39,7 @@ export async function createTemplateApp(ctx) {
 
   try {
     const path = name ? normalize(name) : normalize(template);
+    const keys = generateVAPIDKeys();
     const result = {
       definition: {
         ...template.definition,
@@ -45,6 +47,8 @@ export async function createTemplateApp(ctx) {
         name: name || template,
       },
       private: Boolean(isPrivate),
+      vapidPublicKey: keys.publicKey,
+      vapidPrivateKey: keys.privateKey,
       OrganizationId: organizationId,
       ...(resources && {
         Resources: [].concat(
@@ -70,7 +74,7 @@ export async function createTemplateApp(ctx) {
       result.path = `${path}-${crypto.randomBytes(5).toString('hex')}`;
     }
 
-    const record = await App.create(result, { raw: true, include: [Resource] });
+    const record = await App.create(result, { include: [Resource] });
 
     ctx.body = getAppFromRecord(record);
     ctx.status = 201;