+++ packages/server/migrations/index.js
@@ -1,3 +1,4 @@
 import postgresql from './0.9.0-postgresql';
+import notifications from './0.9.3-notifications';
 
-export default [postgresql];
+export default [postgresql, notifications];