+++ CHANGELOG.md
@@ -10,6 +10,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - App: Add support for absolute URLs in `link` actions.
+- App: Add support for push notifications. Users can subscribe to notifications in the app’s
+  settings page. More strategies for receiving push notifications will be added in the future.
+- App: Add settings page
+- Server: Add support for broadcasting push notifications.
 
 ## [0.9.2] - 2019-11-14